refactor: combine chapter UI
